renamed users file to transactions.js
